Response: allow entire doc to be signed
This is perhaps a bug in php's implementation of Xpath, but
"/blah//selector" is supposed to be:

if (blah.matches(selector) || blah.any_descendent_matches(selector))

But it's not working when blah.matches(selector), so getting rid of
`/blah` works.

Note: the contents of the element has already been verified, we're just
looking for the element that was signed.