Use SimpleXML to insert values into AuthnRequest
to avoid the headache of ensuring that AssertionConsumerServiceURL is properly encoded/escaped for both XML and URL (per XSD standard), using abstracted XML interface such as DOMDocument or SimpleXML handles encoding a given value into well-formed XML value. And if we adopt this approach for one non-static value, we may as well use it for all of them for consistency and future-proofing.