Issue 102 Encryption is enabled in the metadata even though it's disabled
* Pass the boolean and combination of the security settings for
wantNameIdEncrypted and wantAssertionsEncrypted to the
OneLogin_Saml2_Metadata::addX509KeyDescriptors method so the method can
determine whether an encryption KeyDescriptor is necessary. If both settings are
false then an encryption KeyDescriptor will be omitted.
* Updated unit tests to verify that encryption KeyDescriptor is not added if
want_encrypted is not passed as true.
* A potential issue with this change is that since the default values for
wantNameIdEncrypted and wantAssertionsEncrypted are false, if a Service Provider
is setup so that they do not set either value to true and have a certificate set
in their settings the metadata returned by their service provider will change
with this code change. This could be a breaking change in this case.

Issue #102