Fix signature validation issues on Encrypted Assertions when no namespace
Some customers reported that valid encrypted assertions where rejeted,
after some research I discovered that executing directly the xmlsec command
those SAMLResponses were valid, so the issue was related to the way the toolkit
decrypt.

The $objenc->decryptNode call with true parameter decrypt and replace the encrypted
data with the decrypted assertion. The generated XML fails the signature validation
process so I had to replace the $objenc->decryptNode call, use false in order to
get the decrypted content without modifications and store it in order to use it
later on the validation process.