Add support for attribute name mapping and value filtering
At present, the php-saml library reports the attribute names as specified
by the IdP to the caller.  In some cases, it is necessary or useful to
rename, or "map", these names to a more friendly form.  For example, an IdP
may use OIDs in place of "mail" or "eduPersonPrincipalName".  It can also
be useful to filter out known-bad or unwanted values from the returned
attribute values.

This patch adds two new parameters to the SP configuration block:
"attributeMap" and "attributePolicy".  These perform the same basic roles
as the similarly named XML files in the Shibboleth SP software.

The attributeMap parameter is an associative array of
'source'=>'destination' attribute name mappings.  This supports the
merging of many source attributes into a single destination attribute.  If
an attribute name is not present as a key, the attribute is passed through
as-is.

The attributePolicy parameter is an associative array of filtering
functions where the keys are the 'destination' attribute names defined in
attributeMap.  These take a sole argument of an array of values and return
a filtered array of values.  If an attribute name is not present, the
values are passed through unfiltered.